feat: add layout to 404 page, fix middleware rout after logout
