+++ src/widgets/index.ts
@@ -5,3 +5,4 @@ export { HeaderApp } from './header';
 export { FooterApp } from './footer';
 export { AuthWidget } from './auth-widget/auth-widget';
 export { CodeGeneration } from './code-generation';
+export { NotFoundWidget } from './not-found';
